+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # main (unreleased)
 
 - **[FEATURE]**: feat(capture): add support for capture tracepoints [#34](https://github.com/intergral/deep/pull/34) [@Umaaz](https://github.com/Umaaz)
+- **[BUGFIX]**: fix(duration): snapshot duration not set [#37](https://github.com/intergral/deep/pull/37) [@Umaaz](https://github.com/Umaaz)
 
 # 1.1.0 (06/02/2024)